+++ src/main/scala/monocle/functions/Each.scala
@@ -1,0 +1,6 @@
+package monocle.functions
+
+import monocle._
+
+abstract class Each[S, A]:
+  def each: Optic[GetMany & Modify, S, A]
