+++ src/main/scala/monocle/internal/Const.scala
@@ -1,0 +1,3 @@
+package monocle.internal
+
+case class Const[+A, +B](getConst: A)
