+++ src/main/scala/monocle/internal/Proxy.scala
@@ -1,0 +1,7 @@
+package monocle.internal
+
+case class Proxy[+A]()
+
+object Proxy {
+  val nothing: Proxy[Nothing] = Proxy()
+}
