--- src/main/scala/monocly/functions/Each.scala
@@ -1,6 +1,0 @@
-package monocly.functions
-
-import monocly._
-
-abstract class Each[S, A]:
-  def each: Optic[GetMany & Modify, S, A]
