--- src/main/scala/monocly/internal/Const.scala
@@ -1,3 +1,0 @@
-package monocly.internal
-
-case class Const[+A, +B](getConst: A)
