--- src/main/scala/monocly/internal/Proxy.scala
@@ -1,7 +1,0 @@
-package monocly.internal
-
-case class Proxy[+A]()
-
-object Proxy {
-  val nothing: Proxy[Nothing] = Proxy()
-}
